+++ app/serializers/folio/console/file_serializer.rb
@@ -121,4 +121,28 @@ class Folio::Console::FileSerializer
   attribute :additional_html_api_url do |object|
     Rails.application.config.folio_console_files_additional_html_api_url_lambda.call(object)
   end
+
+  attribute :file_modal_additional_fields do |object|
+    object.file_modal_additional_fields.map do |name, hash|
+      h = {
+        name:,
+        type: hash[:type],
+        label: hash[:label] || object.class.human_attribute_name(name),
+        value: object.send(name),
+      }
+
+      if hash[:collection]
+        if hash[:include_blank] != false
+          h[:collection] = [["", ""]]
+        else
+          h[:value] ||= hash[:collection][0][1]
+          h[:collection] = []
+        end
+
+        h[:collection] += hash[:collection]
+      end
+
+      h
+    end
+  end
 end